Made Audio Visualizer it's own control and all use one shared instance
