+++ OsuPlayer/Views/CustomControls/AudioVisualizerView.axaml
@@ -1,0 +1,19 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:lvc="clr-namespace:LiveChartsCore.SkiaSharpView.Avalonia;assembly=LiveChartsCore.SkiaSharpView.Avalonia"
+             xmlns:views="clr-namespace:OsuPlayer.Views.CustomControls"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="OsuPlayer.Views.CustomControls.AudioVisualizerView">
+    <Design.DataContext>
+        <views:AudioVisualizerViewModel />
+    </Design.DataContext>
+
+    <Grid IsHitTestVisible="False">
+        <lvc:CartesianChart IsHitTestVisible="False" VerticalAlignment="Stretch" HorizontalAlignment="Stretch"
+                            Series="{Binding Series}"
+                            XAxes="{Binding XAxes}"
+                            YAxes="{Binding YAxes}" />
+    </Grid>
+</UserControl>
