+++ OsuPlayer/Windows/Miniplayer.axaml
@@ -80,10 +80,7 @@
 
         <Border Background="{DynamicResource AcrylicBaseColor}" Grid.Row="3" />
 
-        <lvc:CartesianChart Grid.Row="3" IsHitTestVisible="False" VerticalAlignment="Stretch" HorizontalAlignment="Stretch"
-                            Series="{Binding Series}"
-                            XAxes="{Binding XAxes}"
-                            YAxes="{Binding YAxes}" />
+        <ContentControl Grid.Row="3" Content="{Binding MainWindowViewModel.AudioVisualizer}"/>
 
         <DockPanel Grid.Row="3" VerticalAlignment="Center" HorizontalAlignment="Center">
             <Panel>
