+++ OsuPlayer/Windows/MiniplayerViewModel.cs
@@ -1,9 +1,5 @@
 ﻿using Avalonia.Media.Imaging;
 using Avalonia.Threading;
-using LiveChartsCore;
-using LiveChartsCore.Defaults;
-using LiveChartsCore.SkiaSharpView;
-using LiveChartsCore.SkiaSharpView.Painting;
 using Nein.Base;
 using Nein.Extensions;
 using OsuPlayer.Data.DataModels.Interfaces;
@@ -14,12 +10,13 @@ using OsuPlayer.IO.Storage.Playlists;
 using OsuPlayer.Modules;
 using OsuPlayer.Modules.Audio.Interfaces;
 using ReactiveUI;
-using SkiaSharp;
 
 namespace OsuPlayer.Windows;
 
 public class MiniplayerViewModel : BaseWindowViewModel
 {
+    public FluentAppWindowViewModel MainWindowViewModel { get; }
+
     private readonly Bindable<bool> _isPlaying = new();
     private readonly Bindable<RepeatMode> _isRepeating = new();
     private readonly Bindable<bool> _isShuffle = new();
@@ -41,57 +38,6 @@ public class MiniplayerViewModel : BaseWindowViewModel
     public bool IsAPlaylistSelected => Player.SelectedPlaylist.Value != default;
 
     public bool IsCurrentSongOnBlacklist => new Blacklist().Contains(CurrentSong.Value);
-
-    #region Audio Visualizer Stuff
-
-    private DispatcherTimer _audioVisualizerUpdateTimer = new();
-
-    public DispatcherTimer AudioVisualizerUpdateTimer
-    {
-        get => _audioVisualizerUpdateTimer;
-        set => this.RaiseAndSetIfChanged(ref _audioVisualizerUpdateTimer, value);
-    }
-
-    private ObservableCollection<ObservableValue> _seriesValues = new();
-
-    public ObservableCollection<ObservableValue> SeriesValues
-    {
-        get => _seriesValues;
-        set => this.RaiseAndSetIfChanged(ref _seriesValues, value);
-    }
-
-    private ISeries[] _series;
-
-    public ISeries[] Series
-    {
-        get => _series;
-        set => this.RaiseAndSetIfChanged(ref _series, value);
-    }
-
-    public Axis[] XAxes { get; set; } =
-    {
-        new Axis
-        {
-            SeparatorsPaint = null,
-            LabelsPaint = null,
-            ShowSeparatorLines = false,
-            MaxLimit = 256
-        }
-    };
-
-    public Axis[] YAxes { get; set; } =
-    {
-        new Axis
-        {
-            MinLimit = 0,
-            MaxLimit = 1,
-            SeparatorsPaint = null,
-            LabelsPaint = null,
-            ShowSeparatorLines = false,
-        }
-    };
-
-    #endregion
 
     public double Volume
     {
@@ -187,9 +133,11 @@ public class MiniplayerViewModel : BaseWindowViewModel
 
     public string ActivePlaylist => $"Active playlist: {Player.SelectedPlaylist.Value?.Name ?? "none"}";
 
-    public MiniplayerViewModel(IPlayer player, IAudioEngine bassEngine)
+    public MiniplayerViewModel(IPlayer player, IAudioEngine bassEngine, FluentAppWindowViewModel mainWindowViewModel)
     {
         Player = player;
+
+        MainWindowViewModel = mainWindowViewModel;
 
         _songTime.BindTo(bassEngine.ChannelPosition);
         _songTime.BindValueChanged(_ => this.RaisePropertyChanged(nameof(SongTime)));
@@ -240,29 +188,5 @@ public class MiniplayerViewModel : BaseWindowViewModel
             this.RaisePropertyChanged(nameof(IsAPlaylistSelected));
             this.RaisePropertyChanged(nameof(IsCurrentSongInPlaylist));
         }, true);
-
-        #region Init Audio Visualizer
-
-        const int size = 4096;
-
-        SeriesValues = new ObservableValue[size].ToObservableCollection();
-
-        for (var i = 0; i < size; i++)
-        {
-            SeriesValues[i] = new ObservableValue(0);
-        }
-
-        Series = new ISeries[]
-        {
-            new ColumnSeries<ObservableValue>
-            {
-                Values = SeriesValues,
-                IsHoverable = false,
-                Fill = new SolidColorPaint(new SKColor(164, 164, 164, 75)),
-                Stroke = null
-            }
-        };
-
-        #endregion
     }
 }
